adding release build support

+++ scripts/server-cron-build.sh
@@ -22,6 +22,11 @@ start_time=`date +%s`
 source `dirname $0`/config.conf
 export QGIS_ANDROID_BUILD_ALL=1
 
+if [[ "$BUILD_TYPE" = "Release" && ! -n "${RELEASE_NAME+x}" ]]; then
+    echo "script aborted, SET: BUILD_TYPE=Release RELEASE_NAME=alpha5 ./server-cron-build.sh";
+    exit 0;
+fi 
+
 cd $QGIS_DIR
 git reset --hard HEAD
 git pull
@@ -31,7 +36,12 @@ git reset --hard HEAD
 git pull
 $SCRIPT_DIR/build-qgis-and-apk.sh
 
-cp -vf $APK_DIR/bin/qgis-debug.apk /home/mbernasocchi/www/download/qgis-nightly.apk
+if [[ "$BUILD_TYPE" = "Release" ]]; then 
+    cp -vf $APK_DIR/bin/qgis-release.apk /home/mbernasocchi/www/download/qgis-$RELEASE_NAME.apk
+    echo $RELEASE_NAME >> /home/mbernasocchi/www/download/versions.txt
+else 
+    cp -vf $APK_DIR/bin/qgis-debug.apk /home/mbernasocchi/www/download/qgis-nightly.apk
+fi
 
 end_time=`date +%s`
 seconds=`expr $end_time - $start_time`
